+++ docker/build.d/shop/810_handle_DEV_TOOLS.sh
@@ -1,6 +1,0 @@
-#!/bin/sh
-
-if [ "$DEV_TOOLS" = "on" ]; then
-  sectionHead "Installing ops tools"
-  install_packages vim less tree
-fi