+++ docker/deploy.d/100_start.sh
@@ -1,0 +1,3 @@
+#!/bin/sh
+
+sectionHead "Starting deployment procedure ..."